requirements.txt
requests==2.31.0
python-telegram-bot==13.15
environs==9.5.0